backend/requirements.txt
Flask==2.3.3
Pillow==11.1.0
python-dotenv==1.0.1
requests==2.31.1  # Updated version
flask-cors==4.0.1  # Updated version
